+++ src/main/java/act/e2e/E2E.java
@@ -25,12 +25,9 @@ import act.app.App;
 import act.app.DbServiceManager;
 import act.db.Dao;
 import act.db.DbService;
-import act.db.sql.tx.TxContext;
 import act.e2e.macro.Macro;
 import act.e2e.req_modifier.RequestModifier;
-import act.e2e.util.RequestTemplateManager;
-import act.e2e.util.ScenarioManager;
-import act.e2e.util.YamlLoader;
+import act.e2e.util.*;
 import act.e2e.verifier.Verifier;
 import act.job.OnAppStart;
 import act.sys.Env;
@@ -40,6 +37,7 @@ import org.osgl.logging.LogManager;
 import org.osgl.logging.Logger;
 import org.osgl.mvc.annotation.DeleteAction;
 import org.osgl.mvc.annotation.PostAction;
+import org.osgl.util.C;
 import org.osgl.util.S;
 
 import java.util.ArrayList;
@@ -49,22 +47,6 @@ import javax.inject.Inject;
 
 @Env.RequireMode(Act.Mode.DEV)
 public class E2E extends LogSupport {
-
-    public enum Status {
-        PENDING, PASS, FAIL;
-
-        public static Status of(boolean result) {
-            return result ? PASS : FAIL;
-        }
-
-        public boolean finished() {
-            return PENDING != this;
-        }
-
-        public boolean pass() {
-            return PASS == this;
-        }
-    }
 
     static final Logger LOGGER = LogManager.get(E2E.class);
 
@@ -123,18 +105,18 @@ public class E2E extends LogSupport {
             List<Dao> list = new ArrayList<>(toBeDeleted);
             for (Dao dao : list) {
                 try {
-                    TxContext.enterTxScope(false);
+                    TxScope.enter();
                     dao.drop();
                     try {
-                        TxContext.exitTxScope();
+                        TxScope.commit();
                     } catch (Exception e) {
                         continue;
                     }
                     toBeDeleted.remove(dao);
                 } catch (Exception e) {
-                    TxContext.exitTxScope(e);
+                    // ignore and try next dao
                 } finally {
-                    TxContext.clear();
+                    TxScope.clear();
                 }
             }
         }
@@ -156,12 +138,13 @@ public class E2E extends LogSupport {
             registerTypeConverters();
             RequestTemplateManager requestTemplateManager = new RequestTemplateManager();
             requestTemplateManager.load();
-            ScenarioManager scenarioManager = new ScenarioManager();
+            final ScenarioManager scenarioManager = new ScenarioManager();
             Map<String, Scenario> scenarios = scenarioManager.load();
             if (scenarios.isEmpty()) {
                 LOGGER.warn("No scenario defined.");
             } else {
-                for (Scenario scenario : scenarios.values()) {
+                C.List<Scenario> list = C.list(scenarios.values()).sorted(new ScenarioComparator(scenarioManager));
+                for (Scenario scenario : list) {
                     scenario.start(scenarioManager, requestTemplateManager);
                 }
             }
@@ -223,7 +206,9 @@ public class E2E extends LogSupport {
             Scenario dep = manager.get(s);
             addToList(dep, list, manager);
         }
-        list.add(scenario);
+        if (!list.contains(scenario)) {
+            list.add(scenario);
+        }
     }
 
 }